Extract prepare_dataset as a standalone script
