+++ src/utils/dataset_utils.py
@@ -1,23 +1,12 @@
 #!/usr/bin/env python3
-import argparse
 import json
 import pandas as pd
-from pathlib import Path
-
-def parse_args():
-    parser = argparse.ArgumentParser(description="Import dataset")
-    parser.add_argument('--train', required=True, type=Path)
-    parser.add_argument('--test', type=Path)
-    parser.add_argument('--class-col', required=True, dest='class_col')
-    parser.add_argument('--description', required=True, type=Path)
-    parser.add_argument('--name', required=True)
-    parser.add_argument('--positive-class')
-    parser.add_argument('--negative-class')
-    parser.add_argument('--output-dir', type=Path, default=Path('datasets'))
-    return parser.parse_args()
 
 def prepare_dataset(train, test, class_col, description, name, 
                    positive_class, negative_class, output_dir):
+    """
+    Preprocesses dataset files to a format digestable by the agent code
+    """
    
     train_df = pd.read_csv(train)
     test_df = pd.read_csv(test) if test else None
@@ -27,16 +16,19 @@ def prepare_dataset(train, test, class_col, description, name,
     all_labels = train_df[label_col].dropna()
     if test_df is not None:
         all_labels = pd.concat([all_labels, test_df[label_col].dropna()])
+    else:
+        print("INFO: No test set found.")
     labels = list(all_labels.unique())
     
     if len(labels) == 2 and positive_class and negative_class:
         # binary classification 
         label_map = {negative_class: 0, positive_class: 1}
-        print(f"Binary mapping: {negative_class}=0, {positive_class}=1")
+        print(f"INFO: Label to number mapping: {label_map}. If this is wrong, please provide positive-class and negative-class parameters to the prepare_dataset script.")
+
     else:
         # multiclass: alphabetical order
         label_map = {lbl: i for i, lbl in enumerate(sorted(labels, key=str))}
-        print(f"Label mapping: {label_map}")
+        print(f"INFO: Label to number mapping: {label_map}")
     
     dataframes = [('train', train_df)]
     if test_df is not None:
@@ -55,8 +47,12 @@ def prepare_dataset(train, test, class_col, description, name,
             out_dir / f'{split_name}.no_label.csv', index=False
         )
     
-    description_content = description.read_text()
-    (out_dir / 'dataset_description.md').write_text(description_content)
+    if(description is not None):
+        description_content = description.read_text()
+        (out_dir / 'dataset_description.md').write_text(description_content)
+    else:
+        print("INFO: No dataset description provided.")
+        (out_dir / 'dataset_description.md').write_text("No dataset description available.")
     
     meta = {
         'label_to_scalar': label_map,
@@ -68,6 +64,9 @@ def prepare_dataset(train, test, class_col, description, name,
 
 def setup_agent_datasets(dataset_dir, dataset_agent_dir):
     dataset_agent_dir.mkdir(parents=True, exist_ok=True)
+    """
+    Links non-sentitive (non-test) dataset files to a directory accessible by agents.
+    """
 
     target_files = ['dataset_description.md', 'train.csv', 'train.no_label.csv']
 
@@ -87,15 +86,3 @@ def setup_agent_datasets(dataset_dir, dataset_agent_dir):
 
                     target_file.symlink_to(source_file)
 
-if __name__ == '__main__':
-    args = parse_args()
-    prepare_dataset(
-        train=args.train,
-        test=args.test,
-        class_col=args.class_col,
-        description=args.description,
-        name=args.name,
-        positive_class=args.positive_class,
-        negative_class=args.negative_class,
-        output_dir=args.output_dir
-    )
